requirements.txt
PyYAML==6.0.1
jinja2==3.1.2
jsonpath-ng==1.6.0
lxml==4.9.3
fastapi==0.104.1
uvicorn==0.24.0.post1
